backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.45.0
google-api-python-client==2.159.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
supabase==2.11.0
apscheduler==3.10.4
python-dotenv==1.0.1
google-cloud-texttospeech==2.21.1
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
